modificacion base de datos y usuario

+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/Mappers.kt
@@ -22,7 +22,7 @@ data class UserUpdateMapper(val email:String,
                             val cvu:String,
                             val walletAddress:String)
 
-data class UserViewMapper(val id: Int?,
+data class UserViewMapper(val id: Long?,
                           val name: String?,
                           val surname: String?,
                           val email:String?,
