+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/service/UserService.kt
@@ -82,8 +82,8 @@ class UserService: UserDetailsService {
     }
 
     @Transactional
-    fun findByID(id: Int): Usuario {
-       val user =  repository.findById(id)
+    fun findByID(id: Long): Usuario {
+       val user =  repository.findById(id.toInt())
        if ( ! (user.isPresent ))
        {throw ItemNotFoundException("User with Id:  $id not found") }
        val newUser=  user.get()
@@ -95,25 +95,25 @@ class UserService: UserDetailsService {
 
 
     @Transactional
-    fun deleteById(id: Int) {
-        val user =  repository.findById(id)
+    fun deleteById(id: Long) {
+        val user =  repository.findById(id.toInt())
         if ( ! (user.isPresent ))
         {throw ItemNotFoundException("User with Id:  $id not found") }
-       repository.deleteById(id)
+       repository.deleteById(id.toInt())
 
     }
 
 
 
     @Transactional
-    fun update(id: Int , entity: UserUpdateMapper) : UserViewMapper {
+    fun update(id: Long, entity: UserUpdateMapper) : UserViewMapper {
        lateinit var   entityOptional:Optional<Usuario?>
         try {
-              entityOptional = repository.findById(id)
+              entityOptional = repository.findById(id.toInt())
              val  newUser:Usuario = entityOptional.get()
             val password =encoder.encode(entity.password)
             val user = Usuario(newUser.id,newUser.name, newUser.surname, entity.email,entity.address,password,entity.cvu,entity.walletAddress)
-            repository.deleteById(id)
+            repository.deleteById(id.toInt())
             val savedUser = repository.save(user)
             val userView =   UserViewMapper(savedUser.id, savedUser.name, savedUser.surname, savedUser.email, savedUser.address, savedUser.cvu, savedUser.walletAddress)
              return userView
